move ui_version feature to new keeper
