+++ public/config.json
@@ -1,0 +1,3 @@
+{
+  "uiVersion": "1.5"
+}
