+++ src/domain/legacy.ts
@@ -16,25 +16,22 @@ import Vault from "sdk/abis/Vault.json";
 
 import { ARBITRUM, AVALANCHE, AVALANCHE_FUJI, getChainName, getConstant, getHighExecutionFee } from "config/chains";
 import { getContract } from "config/contracts";
-import { DECREASE, INCREASE, SWAP, getOrderKey } from "lib/legacy";
 import { USD_DECIMALS } from "config/factors";
+import { DECREASE, INCREASE, SWAP, getOrderKey } from "lib/legacy";
 
 import { t } from "@lingui/macro";
 import { getServerBaseUrl, getServerUrl } from "config/backend";
-import { UI_VERSION, isDevelopment } from "config/env";
-import { REQUIRED_UI_VERSION_KEY } from "config/localStorage";
 import { getTokenBySymbol } from "config/tokens";
+import { bigMath } from "lib/bigmath";
 import { callContract, contractFetcher } from "lib/contracts";
+import { OrderMetricId } from "lib/metrics";
 import { BN_ZERO, bigNumberify, expandDecimals, parseValue } from "lib/numbers";
 import { getProvider, useJsonRpcProvider } from "lib/rpc";
 import { getGmxGraphClient, nissohGraphClient } from "lib/subgraph/clients";
 import groupBy from "lodash/groupBy";
+import useSWRInfinite from "swr/infinite";
 import { replaceNativeTokenAddress } from "./tokens";
 import { getUsd } from "./tokens/utils";
-import useWallet from "lib/wallets/useWallet";
-import useSWRInfinite from "swr/infinite";
-import { bigMath } from "lib/bigmath";
-import { OrderMetricId } from "lib/metrics";
 
 export * from "./prices";
 
@@ -482,29 +479,6 @@ export function useStakedGmxSupply(signer, active) {
   };
 
   return { data, mutate };
-}
-
-export function useHasOutdatedUi() {
-  const { active } = useWallet();
-
-  const url = getServerUrl(ARBITRUM, `/ui_version?client_version=${UI_VERSION}&active=${active}`);
-  const { data, mutate } = useSWR([url], {
-    // @ts-ignore
-    fetcher: (url) => fetch(url).then((res) => res.text()),
-  });
-
-  let hasOutdatedUi = false;
-
-  if (data && parseFloat(data) > parseFloat(UI_VERSION)) {
-    hasOutdatedUi = true;
-  }
-
-  if (isDevelopment()) {
-    const localStorageVersion = localStorage.getItem(REQUIRED_UI_VERSION_KEY);
-    hasOutdatedUi = Boolean(localStorageVersion && parseFloat(localStorageVersion) > parseFloat(UI_VERSION));
-  }
-
-  return { data: hasOutdatedUi, mutate };
 }
 
 export function useGmxPrice(chainId, libraries, active) {
