+++ src/domain/synthetics/tokens/index.ts
@@ -4,5 +4,4 @@ export * from "./useTokenBalances";
 export * from "./useTokenRecentPricesData";
 export * from "./useTokensDataRequest";
 export * from "./use24hPriceDeltaMap";
-export * from "./useOracleKeeperFetcher";
 export * from "./utils";
