+++ src/domain/synthetics/tokens/useTokenRecentPricesData.ts
@@ -5,7 +5,7 @@ import { getToken, getWrappedToken, NATIVE_TOKEN_ADDRESS } from "config/tokens";
 import { LEADERBOARD_PRICES_UPDATE_INTERVAL, PRICES_UPDATE_INTERVAL } from "lib/timeConstants";
 
 import { TokenPricesData } from "./types";
-import { useOracleKeeperFetcher } from "./useOracleKeeperFetcher";
+import { useOracleKeeperFetcher } from "../../../lib/oracleKeeperFetcher/useOracleKeeperFetcher";
 import { useSequentialTimedSWR } from "./useSequentialTimedSWR";
 import { parseContractPrice } from "./utils";
