+++ src/lib/metrics/useConfigureMetrics.ts
@@ -1,7 +1,6 @@
 import { getAbFlags } from "config/ab";
 import { SHOW_DEBUG_VALUES_KEY } from "config/localStorage";
 import { getIsLargeAccount } from "domain/stats/isLargeAccount";
-import { useOracleKeeperFetcher } from "domain/synthetics/tokens";
 import { useChainId } from "lib/chains";
 import { useLocalStorageSerializeKey } from "lib/localStorage";
 import { useBowser } from "lib/useBowser";
@@ -11,6 +10,7 @@ import useWallet from "lib/wallets/useWallet";
 import { useEffect } from "react";
 import { isHomeSite } from "../legacy";
 import { metrics } from "./Metrics";
+import { useOracleKeeperFetcher } from "lib/oracleKeeperFetcher";
 
 export function useConfigureMetrics() {
   const { chainId } = useChainId();
