+++ src/lib/oracleKeeperFetcher/index.ts
@@ -1,2 +1,3 @@
 export * from "./oracleKeeperFetcher";
+export * from "./useOracleKeeperFetcher";
 export * from "./types";
